Switch IR to sea-of-nodes (sorta)
Conversation with beach a while back made me realize the control
flow among cars and cdrs and such is actually not relevant, at
least not for most optimizations.

What I'd like to go for is a propagation phase that does not
modify the IR at all, or does so locally. That should facilitate
concurrent operation, and also just be... simpler. Less chaotic.
Sea of nodes should help with this, I think? Each worker can work
on and modify a continuation at a time, and add and remove nodes
as it needs.