Real "reoptimize-until-done" loop
Does a lot of redundant work for now, but seems to work. Although
also it's not redundant enough, in that flow will stop before new
instructions unless there's a new data dependency (as
pre-transform instructions will have already propagated). But now
that instruction mapping is cheap this shouldn't be a big problem
to fix by inserting a map-everything pass.