quickc: inline $vau
This works, but I am very rapidly running up against the
limitations of the single pass compiler. I was going to do local
calls next, but we'd have to compute the closure first and then
drop it - and what the hell is the point of that? So next on the
agenda I'm going to move out the infos (not sure how to reconcile
them with flowc yet) and compilation environments, and then start
on treec, which I expect could actually have decent performance.