Flesh out the standard environment a little
Ideally Burke will eventually be self-hosting, but for now it is
not, and the best way to make it self-hosting is to write the
compiler in Burke and set it to work on implementation code.
And for that we need a functioning language.
I'm pretty against the approach in the report of defining things
one at a time. Primitive bootstrapping. No thanks.