Mingle baselib definitions with compilation env definitions
This seems like a much better organization than putting
everything in the compilation-environment file. The downside, such
as it is, is that compilation is more integrated; but honestly
that's pretty much the whole point of all of this, so that's fine.

Don't think I like just setting a compilation-environment
variable in the REPL though. I am considering a subclass of
ENVIRONMENT that just has a compilation environment integrated,
but I think I should work out module compilation in detail first.