Redesign VM object system
My endgoal here is to come up with a VM that is generally useful
for a variety of things, a little like the JVM - but with stuff
that's good for lisps, like nonlocal exits, easy closures, etc.

This change will increase the size of the bytecode, but frees up a
lot of opcodes. More importantly, I hope that it will separate
representation details out in such a way that facilitates
implementing the VM in other languages/systems.
We shall see how that goes.